Combine ink files to mod, fix benches
