+++ src/ink.rs
@@ -5,8 +5,10 @@ use super::drink::{
 };
 use super::drink_api::{CallArgs, ContractsBalanceOf, DrinkApi};
 use crate::drink_api::CreateArgs;
-use crate::ink_build;
-use contract_build::Target;
+use contract_build::{
+    BuildArtifacts, BuildMode, ExecuteArgs, Features, ManifestPath, Network, OptimizationPasses,
+    OutputType, Target, UnstableFlags, Verbosity,
+};
 use ink::{
     codegen::ContractCallBuilder,
     env::{
@@ -19,7 +21,7 @@ use ink::{
 };
 use parity_scale_codec::{Decode, Encode};
 use std::marker::PhantomData;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use subxt_signer::sr25519::{dev, Keypair};
 
 pub struct InkDrink<E: Environment, Runtime: RuntimeT> {
@@ -63,8 +65,7 @@ where
         } else {
             panic!("No VM target feature enabled")
         };
-        let build_result =
-            ink_build::build_contract(contract, target).expect("Error building contract");
+        let build_result = build_contract(contract, target).expect("Error building contract");
         let code = std::fs::read(build_result).expect("Error loading contract");
 
         let caller = dev::alice();
@@ -143,3 +144,41 @@ pub type CallBuilderFinal<E, Args, RetType> = ink::env::call::CallBuilder<
     Set<ExecutionInput<Args>>,
     Set<ReturnType<RetType>>,
 >;
+
+/// Builds the contract at `manifest_path`, returns the path to the contract
+/// Wasm build artifact.
+pub fn build_contract<P>(path_to_cargo_toml: P, target: Target) -> anyhow::Result<PathBuf>
+where
+    P: AsRef<Path> + Copy,
+{
+    let manifest_path = ManifestPath::new(path_to_cargo_toml).unwrap_or_else(|err| {
+        panic!(
+            "Invalid manifest path {}: {err}",
+            path_to_cargo_toml.as_ref().display()
+        )
+    });
+    let args = ExecuteArgs {
+        manifest_path,
+        verbosity: Verbosity::Default,
+        build_mode: BuildMode::Release,
+        features: Features::default(),
+        network: Network::Online,
+        build_artifact: BuildArtifacts::CodeOnly,
+        unstable_flags: UnstableFlags::default(),
+        optimization_passes: Some(OptimizationPasses::default()),
+        keep_debug_symbols: false,
+        output_type: OutputType::HumanReadable,
+        skip_wasm_validation: false,
+        target,
+        ..Default::default()
+    };
+
+    let build_result = contract_build::execute(args)?;
+
+    let code_artifact_path = build_result
+        .dest_wasm
+        .expect("Wasm code artifact not generated")
+        .canonicalize()?;
+
+    Ok(code_artifact_path)
+}
